backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-dotenv==1.1.0
Pillow==11.2.1
